requirements.txt
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
Flask==2.2.3
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pip==23.0.1
python-telegram-bot==20.1
rfc3986==1.5.0
setuptools==67.4.0
sniffio==1.3.0
typing_extensions==4.4.0
Werkzeug==2.2.3
wheel==0.38.4
zipp==3.15.0
